Add feature to attach all pdfs in given directory
The `.doirc` file can now contain a `pdfpath` argument specifying a path
where all pdf files in the `.bib` file can be found.

The script will then search for pdfs in the directory with the exact bib
entry key. For example, for an entry `authorDateTitle`, the script
expects to find `authorDateTitle.pdf` in the directory tree given by
`pdfpath`.